refactor: 메서드 추출, 패키지 정리
