+++ src/main/java/com/khu/gitbox/domain/file/application/FileService.java
@@ -12,9 +12,9 @@ import com.khu.gitbox.domain.file.entity.File;
 import com.khu.gitbox.domain.file.entity.FileStatus;
 import com.khu.gitbox.domain.file.entity.FileType;
 import com.khu.gitbox.domain.file.infrastructure.FileRepository;
-import com.khu.gitbox.domain.file.presentation.dto.FileCreateRequest;
-import com.khu.gitbox.domain.file.presentation.dto.PullRequestCreateRequest;
+import com.khu.gitbox.domain.file.presentation.dto.request.FileCreateRequest;
 import com.khu.gitbox.domain.file.presentation.dto.request.FileUpdateRequest;
+import com.khu.gitbox.domain.file.presentation.dto.request.PullRequestCreateRequest;
 import com.khu.gitbox.domain.file.presentation.dto.response.FileGetResponse;
 import com.khu.gitbox.domain.member.entity.Member;
 import com.khu.gitbox.domain.member.infrastructure.MemberRepository;
@@ -36,11 +36,11 @@ public class FileService {
 	private final FileRepository fileRepository;
 	private final MemberRepository memberRepository;
 	private final WorkspaceRepository workspaceRepository;
-	private final S3Service s3Service;
 	private final PullRequestRepository pullRequestRepository;
+	private final S3Service s3Service;
 
 	// 파일 업로드
-	public Long uploadFile(FileCreateRequest request, MultipartFile multipartFile) {
+	public FileGetResponse uploadFile(FileCreateRequest request, MultipartFile multipartFile) {
 		final Member member = getCurrentMember();
 		final Workspace workspace = getAvailableWorkspace(request.workspaceId());
 
@@ -61,16 +61,16 @@ public class FileService {
 			.rootFileId(null)
 			.parentFileId(null)
 			.build();
-
 		workspace.increaseUsedStorage(file.getSize());
 
 		final File savedFile = fileRepository.save(file);
 		savedFile.updateRootFileId(savedFile.getId());
-		return fileRepository.save(savedFile).getId();
+
+		return FileGetResponse.of(fileRepository.save(savedFile));
 	}
 
 	// 새로운 버전 파일 업로드 (+ PR 생성)
-	public Long uploadNewVersionFile(
+	public FileGetResponse uploadNewVersionFile(
 		Long parentFileId,
 		PullRequestCreateRequest request,
 		MultipartFile multipartFile) {
@@ -78,16 +78,8 @@ public class FileService {
 		final Member member = getCurrentMember();
 		final File parentFile = getAvailableFile(parentFileId);
 
-		// 부모 파일이 최신 버전인지 확인
-		if (!parentFile.isLatest()) {
-			throw new CustomException(HttpStatus.BAD_REQUEST, "부모 파일이 최신 버전이 아닙니다.");
-		}
-
-		// 이미 업데이트 대기 중인 파일이 없는지 확인 (PR 여부)
-		fileRepository.findPendingFile(parentFileId)
-			.ifPresent(pendingFile -> {
-				throw new CustomException(HttpStatus.BAD_REQUEST, "이미 업데이트 대기 중인 파일이 있습니다.");
-			});
+		// 부모 파일 업데이트 가능 여부 확인
+		validateParentFileForUpdate(parentFileId, parentFile);
 
 		// 파일 업로드 (PR 승인 시 부모 파일을 구버전으로)
 		final String fileName = multipartFile.getOriginalFilename();
@@ -122,7 +114,7 @@ public class FileService {
 		final Workspace workspace = getAvailableWorkspace(parentFile.getWorkspaceId());
 		workspace.increaseUsedStorage(savedFile.getSize());
 
-		return savedFile.getId();
+		return FileGetResponse.of(savedFile);
 	}
 
 	// 파일 조회
@@ -178,4 +170,16 @@ public class FileService {
 		return fileName.substring(fileName.lastIndexOf(".") + 1);
 	}
 
+	private void validateParentFileForUpdate(Long parentFileId, File parentFile) {
+		// 부모 파일이 최신 버전인지 확인
+		if (!parentFile.isLatest()) {
+			throw new CustomException(HttpStatus.BAD_REQUEST, "부모 파일이 최신 버전이 아닙니다.");
+		}
+
+		// 이미 업데이트 대기 중인 파일이 없는지 확인 (PR 여부)
+		fileRepository.findPendingFile(parentFileId)
+			.ifPresent(pendingFile -> {
+				throw new CustomException(HttpStatus.BAD_REQUEST, "이미 업데이트 대기 중인 파일이 있습니다.");
+			});
+	}
 }
