+++ src/main/java/com/khu/gitbox/domain/file/presentation/dto/request/PullRequestCreateRequest.java
@@ -1,4 +1,4 @@
-package com.khu.gitbox.domain.file.presentation.dto;
+package com.khu.gitbox.domain.file.presentation.dto.request;
 
 public record PullRequestCreateRequest(
 	String title,
